requirements.txt
# Django y DRF
Django~=4.2.23
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0

# Base de datos
psycopg2-binary~=2.9.9

# HTTP Requests
requests~=2.32.4

# Tareas asíncronas
celery~=5.3.4
redis~=5.0.1
django-celery-beat~=2.5.0

# Documentación API
drf-yasg~=1.21.7

# Testing
pytest~=7.4.3
pytest-django~=4.7.0
pytest-cov~=4.1.0

# Linting y formateo
black~=23.12.1
isort~=5.13.2
flake8~=6.1.0

# Variables de entorno
python-decouple~=3.8 